Comment deletion Form and Alert Notification after delete

+++ app/Http/Controllers/CommentsController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use App\Comment;
 use App\Post;
 use App\Http\Requests\CommentsRequest;
+use Illuminate\Auth\Access\AuthorizationException;
 use Illuminate\Http\Request;
 use Illuminate\Http\Response;
 use Illuminate\Support\Facades\Auth;
@@ -39,16 +40,17 @@ class CommentsController extends Controller {
 //        //
 //    }
 
-    /**
-     * Remove the specified resource from storage.
-     *
-     * @param Comment $comment
-     * @return Response
-     */
-    public function destroy(Comment $id) {
-		$comment = Post::findOrFail($id);
-		$post = $comment->post();
+	/**
+	 * Remove the specified resource from storage.
+	 *
+	 * @param Comment $comment
+	 * @return Response
+	 * @throws AuthorizationException
+	 * @throws \Exception
+	 */
+    public function destroy(Comment $comment) {
+		$this->authorize('delete', $comment);
 		$comment->delete();
-		return redirect()->route('posts.show', $post)->with('success', "comment deleted");
+		return redirect()->route('posts.show', $comment->post_id)->with('success', "Comment Deleted");
 	}
 }
